[Jobs] Split jobs in several files
Split the `jobs.py` file and put each separate job in its own
file. This simplifies the structure of the code.

Renamed the subprocesses for pending messages and txs to make
them similar.
